refactor presentors for logincode add TimetablePresenter to move presenter methods from ControllerInputTimetable
