+++ pythonfilesDONOTMARK/legacySwitch.py
@@ -8,42 +8,42 @@ import pyperclip
 
 INPUT = R"""
             case BAN -> {  
-                String userToBan = presenter.enterUsername();  
+                String userToBan = loginPresenter.enterUsername();
                 Date unbanDate = new Date();  
                 try {  
-                    unbanDate = presenter.enterDate();  
+                    unbanDate = loginPresenter.enterDate();
                 } catch (ParseException e) {  
-                    // Unsure if this is supposed to be printed to screen; put into presenter for now.  
+                    // Unsure if this is supposed to be printed to screen; put into loginPresenter for now.
                     // System.out.println("Failed to parse string to date");  
-                    presenter.parseFailure();  
+                    loginPresenter.parseFailure();
                 }  
                 boolean banCheck = adminAccountManager.banUser(userToBan, unbanDate);  
-                presenter.banUserConfirm(banCheck, userToBan);  
+                loginPresenter.banUserConfirm(banCheck, userToBan);
   
                 // banUser(userToBan, unbanDate);  
             }  
   
             case DELETE -> {  
-                String userToDelete = presenter.enterUsername();  
+                String userToDelete = loginPresenter.enterUsername();
                 boolean delUserSuccess = adminAccountManager.deleteUser(userToDelete);  
-                presenter.deleteUserConfirm(delUserSuccess, userToDelete);  
+                loginPresenter.deleteUserConfirm(delUserSuccess, userToDelete);
   
                 // deleteUser(userToDelete);  
             }  
   
             case NEW -> {  
-                String[] inputs = presenter.enterCredentials();  
+                String[] inputs = loginPresenter.enterCredentials();
                 boolean isCreated = adminAccountManager.createNewAdminUser(inputs[0], inputs[1]);  
-                presenter.createNewAdminConfirm(isCreated, inputs[0]);  
+                loginPresenter.createNewAdminConfirm(isCreated, inputs[0]);
             }  
   
             case PROMOTE -> {  
-                String userToPromote = presenter.enterUsername();  
-                presenter.promoteUserConfirm(adminAccountManager.addPermission(userToPromote, "admin"), userToPromote);  
+                String userToPromote = loginPresenter.enterUsername();
+                loginPresenter.promoteUserConfirm(adminAccountManager.addPermission(userToPromote, "admin"), userToPromote);
                 // promoteUserToAdmin(userToPromote);  
             }  
             case BACK -> curState = LoggedInState.STANDARD;default -> {  
-                presenter.genericFailedAction("invalid");  
+                loginPresenter.genericFailedAction("invalid");
                 return false;  
             }  
 """.strip()
