+++ src/main/java/org/example/logincode/interfaceadapters/Main.java
@@ -12,34 +12,34 @@ public class Main {
 
     private void runProgram() {
         Controller controller = new Controller();
-        Presenter presenter = new Presenter();
+        LoginPresenter loginPresenter = new LoginPresenter();
         Scanner scanner = new Scanner(System.in);
 
-        login(controller, presenter, scanner);
+        login(controller, loginPresenter, scanner);
 
         String tLine2;
-        tLine2 = controller.presenter.printAndAskPrompt(controller.getCommandList());
+        tLine2 = controller.loginPresenter.printAndAskPrompt(controller.getCommandList());
 
         eventLoop(controller, tLine2);
 
-        presenter.exitProgram();
+        loginPresenter.exitProgram();
         System.exit(0);
     }
 
     private void eventLoop(Controller controller, String tLine2) {
         while (!tLine2.equals("exit")) {
             controller.inputParser(tLine2);
-            tLine2 = controller.presenter.printAndAskPrompt(controller.getCommandList());
+            tLine2 = controller.loginPresenter.printAndAskPrompt(controller.getCommandList());
         }
     }
 
-    private void login(Controller controller, Presenter presenter, Scanner scanner) {
+    private void login(Controller controller, LoginPresenter loginPresenter, Scanner scanner) {
         String line;
         while (controller.getLoginState() == Controller.LoginState.LOGGED_OUT) {
-            presenter.startView();
+            loginPresenter.startView();
             line = scanner.nextLine();
             if (line.equals("exit")) {
-                presenter.exitProgram();
+                loginPresenter.exitProgram();
                 System.exit(0);
             }
             controller.loggedOutMenu(line);
