+++ src/main/java/org/example/logincode/interfaceadapters/controllerinput/ControllerInputTimetable.java
@@ -1,13 +1,11 @@
 package org.example.logincode.interfaceadapters.controllerinput;
 
-import org.example.PresenterPrinter;
 import org.example.coursegetter.entities.Course;
 import org.example.coursegetter.usecases.CourseSearcherGetter;
-import org.example.coursegetter.usecases.CourseSearcherIndividual;
-import org.example.logincode.interfaceadapters.Presenter;
+import org.example.logincode.interfaceadapters.LoginPresenter;
+import org.example.logincode.interfaceadapters.TimetablePresenter;
 import org.example.logincode.usecases.AccountManager;
 import org.example.logincode.usecases.StorageManager;
-import org.example.studentdata.entities.CourseChoice;
 import org.example.studentdata.usecases.StudentManager;
 import org.example.studentdata.usecases.StudentManagerBuilder;
 import org.example.timetable.interfaceadapters.TimetableController;
@@ -17,7 +15,8 @@ import org.example.timetable.usecases.TimetableCommunicatorBulkBuilder;
 public class ControllerInputTimetable extends ControllerInput {
     private final CourseSearcherGetter csg;
     private StudentManager sm;
-    private PresenterPrinter prt = new PresenterPrinter();
+
+    private TimetablePresenter timetablePresenter = new TimetablePresenter();
     private TimetableController ttc;
     private TimetableCommunicatorBulkBuilder timetableCommunicatorBulkBuilder = new TimetableCommunicatorBulkBuilder();
 
@@ -27,11 +26,11 @@ public class ControllerInputTimetable extends ControllerInput {
      *
      * @param manager               always the same manager in the controller class
      * @param accountStorageManager ^
-     * @param presenter             ^
+     * @param loginPresenter             ^
      */
-    public ControllerInputTimetable(AccountManager manager, StorageManager accountStorageManager, Presenter presenter,
+    public ControllerInputTimetable(AccountManager manager, StorageManager accountStorageManager, LoginPresenter loginPresenter,
                                     CourseSearcherGetter csg) {
-        super(manager, accountStorageManager, presenter);
+        super(manager, accountStorageManager, loginPresenter);
         ttc = new TimetableController(getTCB());
         StudentManagerBuilder smb = new StudentManagerBuilder();
         sm = smb.buildStudentManager(manager);
@@ -47,53 +46,53 @@ public class ControllerInputTimetable extends ControllerInput {
                 return true;
             }
             case "addcourse" -> {
-                String rqc = prt.askWithMessage("What course would you like to add?");
+                String rqc = timetablePresenter.addCourse();
                 Course temp = csg.getCourseSearcher().getCourseOfferingByCode("20229", rqc);
                 boolean addedCourseState = sm.addBlankPlannedCourse(temp);
                 if(addedCourseState){
-                    prt.println("Added the course.");
+                    timetablePresenter.addCourseConfirmation();
                 } else {
-                    prt.println("Couldn't add the course - it either didn't exist or was a duplicate.");
+                    timetablePresenter.addCourseError();
                 }
             }
             case "addmeetingtocourse" -> {
-                String courseCode = prt.askWithMessage("What course would you add a section to?");
+                String courseCode = timetablePresenter.addMeetingToCourse();
                 if (sm.getPlannedCourseByString(courseCode) != null) {
-                    String sectionCode = prt.askWithMessage("What section would you like to add?");
+                    String sectionCode = timetablePresenter.addSectionToCourse();
                     boolean setCourseState = sm.setCourseChoice(sm.getPlannedCourseByString(courseCode), sectionCode);
                     if(setCourseState) {
-                        prt.println("Meeting added.");
-                    } else prt.println("That course doesn't have this meeting.");
+                        timetablePresenter.addMeetingConfirmation();
+                    } else timetablePresenter.addMeetingError();
                 }
                 return true;
             }
             case "addprevcourse" -> {
-                String courseCode = prt.askWithMessage("What course would you like to add?");
-                String session = prt.askWithMessage("In which session did you take this course?"); // instructions for year + 9 / 5 etc.?
+                String courseCode = timetablePresenter.addPrevCourse();
+                String session = timetablePresenter.addPrevSessionCourse(); // instructions for year + 9 / 5 etc.?
                 Course temp = csg.getCourseSearcher().getCourseOfferingByCode(session, courseCode);
                 boolean setCourseState = sm.addPreviousCourse(temp);
                 if (setCourseState){
-                    prt.println("Course added.");
-                }else prt.println("That course is not available in that session.");
+                    timetablePresenter.addPrevCourseConfirmation(session);
+                }else timetablePresenter.addPrevCourseError();
 
                 return setCourseState;
             }
             case "delcourse" -> {
-                String courseCode = prt.askWithMessage("What course would you like to delete?");
+                String courseCode = timetablePresenter.deleteCourse();
 
                 boolean setCourseState = sm.removePlannedCourse(sm.getPlannedCourseByString(courseCode));
                 if (setCourseState){
-                    prt.println("Course removed.");
-                }else prt.println("You do not currently have that course.");
+                    timetablePresenter.deleteCourseConfirmation();
+                }else timetablePresenter.deleteCourseError();
 
                 return setCourseState;
             }
             case "delprevcourse" -> {
-                String courseCode = prt.askWithMessage("What previous course would you like to remove?");
+                String courseCode = timetablePresenter.deletePrevCourse();
                 boolean setCourseState = sm.removePreviousCourse(sm.getPreviousCourseByString(courseCode));
                 if (setCourseState){
-                    prt.println("Course removed.");
-                }else prt.println("You have not enrolled in that course before.");
+                    timetablePresenter.deletePrevCourseConfirmation();
+                }else timetablePresenter.deletePrevCourseError();
 
                 return setCourseState;
             }
